requirements.txt
Pillow==8.1.0
pkg-resources==0.0.0
